rename some logical processors
