+++ snuba/datasets/cdc/groupedmessage_entity.py
@@ -9,7 +9,9 @@ from snuba.pipeline.simple_pipeline import SimplePipelineBuilder
 from snuba.query.data_source.join import JoinRelationship, JoinType
 from snuba.query.processors import QueryProcessor
 from snuba.query.processors.logical.basic_functions import BasicFunctionsProcessor
-from snuba.query.processors.object_id_rate_limiter import ProjectRateLimiterProcessor
+from snuba.query.processors.logical.object_id_rate_limiter import (
+    ProjectRateLimiterProcessor,
+)
 
 
 class GroupedMessageEntity(Entity):
