+++ snuba/datasets/entities/events.py
@@ -28,14 +28,14 @@ from snuba.query.logical import Query
 from snuba.query.processors import QueryProcessor
 from snuba.query.processors.handled_functions import HandledFunctionsProcessor
 from snuba.query.processors.logical.basic_functions import BasicFunctionsProcessor
-from snuba.query.processors.object_id_rate_limiter import (
+from snuba.query.processors.logical.object_id_rate_limiter import (
     ProjectRateLimiterProcessor,
     ProjectReferrerRateLimiter,
     ReferrerRateLimiterProcessor,
 )
+from snuba.query.processors.logical.tags_expander import TagsExpanderProcessor
+from snuba.query.processors.logical.timeseries_processor import TimeSeriesProcessor
 from snuba.query.processors.quota_processor import ResourceQuotaProcessor
-from snuba.query.processors.tags_expander import TagsExpanderProcessor
-from snuba.query.processors.timeseries_processor import TimeSeriesProcessor
 from snuba.query.query_settings import QuerySettings
 from snuba.query.validation.validators import EntityRequiredColumnValidator
