+++ tests/query/processors/test_project_referrer_rate_limiter.py
@@ -8,7 +8,9 @@ from snuba.query.conditions import ConditionFunctions, binary_condition
 from snuba.query.data_source.simple import Entity as QueryEntity
 from snuba.query.expressions import Column, Expression, Literal
 from snuba.query.logical import Query
-from snuba.query.processors.object_id_rate_limiter import ProjectReferrerRateLimiter
+from snuba.query.processors.logical.object_id_rate_limiter import (
+    ProjectReferrerRateLimiter,
+)
 from snuba.query.query_settings import HTTPQuerySettings
 from snuba.state.rate_limit import PROJECT_REFERRER_RATE_LIMIT_NAME
