+++ tests/query/processors/test_referrer_rate_limiter.py
@@ -6,7 +6,9 @@ from snuba.query.conditions import ConditionFunctions, binary_condition
 from snuba.query.data_source.simple import Entity as QueryEntity
 from snuba.query.expressions import Column, FunctionCall, Literal
 from snuba.query.logical import Query
-from snuba.query.processors.object_id_rate_limiter import ReferrerRateLimiterProcessor
+from snuba.query.processors.logical.object_id_rate_limiter import (
+    ReferrerRateLimiterProcessor,
+)
 from snuba.query.query_settings import HTTPQuerySettings
 from snuba.state.rate_limit import REFERRER_RATE_LIMIT_NAME
