+++ tests/query/processors/test_timeseries_processor.py
@@ -17,7 +17,7 @@ from snuba.query.dsl import multiply
 from snuba.query.exceptions import InvalidQueryException
 from snuba.query.expressions import Column, FunctionCall, Literal
 from snuba.query.logical import Query
-from snuba.query.processors.timeseries_processor import (
+from snuba.query.processors.logical.timeseries_processor import (
     TimeSeriesProcessor,
     extract_granularity_from_query,
 )
